Cleanup iteration #2

+++ src/main/java/org/socialsignin/spring/data/dynamodb/repository/query/PartTreeDynamoDBQuery.java
@@ -96,7 +96,7 @@ public class PartTreeDynamoDBQuery<T, ID> extends AbstractDynamoDBQuery<T, ID> i
     @Override
     protected boolean isSingleEntityResultsRestriction() {
         Integer resultsRestiction = getResultsRestrictionIfApplicable();
-        return resultsRestiction != null && resultsRestiction.intValue() == 1;
+        return resultsRestiction != null && resultsRestiction == 1;
     }
 
 }
